# openapi-schema.json
openapi: 3.0.3
info:
  title: ChargeGhar Nepal API
  version: a315307-dirty
  description: ChargeGhar API - Shared Power Bank Network for Nepal. This documentation
    only shows endpoints documented in our Features TOC.
  contact:
    name: ChargeGhar API Support
    email: support@chargegh.com
  license:
    name: Proprietary
paths:
...
  /api/app/countries:
    get:
      operationId: api_app_countries_list
      description: Returns a list of countries with dialing codes (e.g., +977 for
        Nepal).
      summary: Get Country Codes
      tags:
      - App
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryList'
          description: ''
  /api/app/countries/search:
    get:
      operationId: api_app_countries_search_retrieve
      description: Search countries by name or code
      tags:
      - api
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryList'
          description: ''
...
  /api/app/init-data:
    get:
      operationId: api_app_init_data_retrieve
      description: Get app initialization data
      tags:
      - api
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInitData'
          description: ''
  /api/app/media/upload:
    post:
      operationId: api_app_media_upload_create
      description: Upload media files like images, videos to Cloudinary. Returns secure
        URLs for storage and CDN access.
      summary: Upload Media
      tags:
      - App
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MediaUploadCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MediaUploadCreateRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadCreate'
          description: ''
  /api/app/media/uploads:
    get:
      operationId: api_app_media_uploads_list
      description: Get user's uploaded media files
      tags:
      - api
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaUpload'
          description: ''
  /api/app/media/uploads/{upload_id}:
    delete:
      operationId: api_app_media_uploads_destroy
      description: Delete a specific media upload by ID
      summary: Delete Media Upload
      parameters:
      - in: path
        name: upload_id
        schema:
          type: string
        description: Media upload ID
        required: true
      tags:
      - App
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    AppInitData:
      type: object
      description: Serializer for app initialization data
      properties:
        configs:
          type: object
          additionalProperties: {}
        countries:
          type: array
          items: {}
        constants:
          type: object
          additionalProperties: {}
        app_version:
          type: string
        api_version:
          type: string
      required:
      - api_version
      - app_version
      - configs
      - constants
      - countries
...
    CountryList:
      type: object
      description: Minimal serializer for country listing
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 100
        code:
          type: string
          maxLength: 10
        dial_code:
          type: string
          maxLength: 10
        flag_url:
          type: string
          format: uri
          nullable: true
          maxLength: 200
      required:
      - code
      - dial_code
      - id
      - name
...
    MediaUpload:
      type: object
      description: Serializer for MediaUpload model
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        file_url:
          type: string
          format: uri
          maxLength: 200
        file_type:
          enum:
          - IMAGE
          - VIDEO
          - DOCUMENT
          type: string
          description: |-
            * `IMAGE` - Image
            * `VIDEO` - Video
            * `DOCUMENT` - Document
          x-spec-enum-id: 7ef6abb5f6ca0d8f
        original_name:
          type: string
          maxLength: 255
        file_size:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        uploaded_by:
          type: integer
          readOnly: true
          nullable: true
        cloud_provider:
          type: string
          readOnly: true
        public_id:
          type: string
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - cloud_provider
      - created_at
      - file_size
      - file_type
      - file_url
      - id
      - original_name
      - public_id
      - updated_at
      - uploaded_by
    MediaUploadCreate:
      type: object
      description: Serializer for creating media uploads
      properties:
        file:
          type: string
          format: uri
          pattern: (?:jpg|jpeg|png|gif|mp4|mov|pdf|doc|docx)$
        file_type:
          enum:
          - IMAGE
          - VIDEO
          - DOCUMENT
          type: string
          description: |-
            * `IMAGE` - Image
            * `VIDEO` - Video
            * `DOCUMENT` - Document
          x-spec-enum-id: 7ef6abb5f6ca0d8f
      required:
      - file
      - file_type
    MediaUploadCreateRequest:
      type: object
      description: Serializer for creating media uploads
      properties:
        file:
          type: string
          format: binary
          pattern: (?:jpg|jpeg|png|gif|mp4|mov|pdf|doc|docx)$
        file_type:
          enum:
          - IMAGE
          - VIDEO
          - DOCUMENT
          type: string
          description: |-
            * `IMAGE` - Image
            * `VIDEO` - Video
            * `DOCUMENT` - Document
          x-spec-enum-id: 7ef6abb5f6ca0d8f
      required:
      - file
      - file_type
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...